requirements.txt
fastapi==0.128.0
discord.py==2.4.0
google-genai==1.59.0
google-search-results==2.4.2
python-dotenv==1.2.1
requests==2.32.5
pynacl==1.6.2
anyio==4.12.1